Move GitHub clients into core
