+++ starter-api/src/main/java/io/micronaut/starter/api/create/github/GitHubCreateController.java
@@ -30,8 +30,8 @@ import io.micronaut.scheduling.TaskExecutors;
 import io.micronaut.scheduling.annotation.ExecuteOn;
 import io.micronaut.starter.api.RequestInfo;
 import io.micronaut.starter.api.TestFramework;
-import io.micronaut.starter.api.create.github.client.v3.GitHubRepository;
 import io.micronaut.starter.application.ApplicationType;
+import io.micronaut.starter.client.github.v3.GitHubRepository;
 import io.micronaut.starter.options.BuildTool;
 import io.micronaut.starter.options.JdkVersion;
 import io.micronaut.starter.options.Language;
