+++ starter-api/src/main/java/io/micronaut/starter/api/create/github/GitHubCreateService.java
@@ -19,16 +19,15 @@ import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.context.event.ApplicationEventPublisher;
-import io.micronaut.core.util.StringUtils;
 import io.micronaut.http.client.exceptions.HttpClientResponseException;
 import io.micronaut.starter.api.StarterConfiguration;
 import io.micronaut.starter.api.TestFramework;
 import io.micronaut.starter.api.create.AbstractCreateController;
-import io.micronaut.starter.api.create.github.client.oauth.AccessToken;
-import io.micronaut.starter.api.create.github.client.oauth.GitHubOAuthClient;
-import io.micronaut.starter.api.create.github.client.v3.GitHubApiClient;
-import io.micronaut.starter.api.create.github.client.v3.GitHubRepository;
-import io.micronaut.starter.api.create.github.client.v3.GitHubUser;
+import io.micronaut.starter.client.github.oauth.AccessToken;
+import io.micronaut.starter.client.github.oauth.GitHubOAuthClient;
+import io.micronaut.starter.client.github.v3.GitHubApiClient;
+import io.micronaut.starter.client.github.v3.GitHubRepository;
+import io.micronaut.starter.client.github.v3.GitHubUser;
 import io.micronaut.starter.application.ApplicationType;
 import io.micronaut.starter.application.generator.GeneratorContext;
 import io.micronaut.starter.application.generator.ProjectGenerator;
@@ -38,11 +37,7 @@ import io.micronaut.starter.io.OutputHandler;
 import io.micronaut.starter.options.BuildTool;
 import io.micronaut.starter.options.JdkVersion;
 import io.micronaut.starter.options.Language;
-import org.eclipse.jgit.api.Git;
-import org.eclipse.jgit.api.errors.GitAPIException;
-import org.eclipse.jgit.transport.PushResult;
-import org.eclipse.jgit.transport.RemoteRefUpdate;
-import org.eclipse.jgit.transport.UsernamePasswordCredentialsProvider;
+import io.micronaut.starter.util.GitHubUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -55,8 +50,6 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
 
 /**
  * GitHub create service.
@@ -68,7 +61,6 @@ import java.util.stream.StreamSupport;
 @Requires(beans = GitHubOAuthClient.class)
 public class GitHubCreateService extends AbstractCreateController {
 
-    protected static final String COMMIT_MESSAGE = "Initial commit";
     private static final Logger LOG = LoggerFactory.getLogger(GitHubCreateService.class);
     private static final String TOKEN_PREFIX = "token ";
     private static final String REPO_PREFIX = "generated";
@@ -124,7 +116,8 @@ public class GitHubCreateService extends AbstractCreateController {
         try {
             repoPath = Files.createTempDirectory(Paths.get(TMP_DIR), REPO_PREFIX);
             generateAppLocally(generatorContext, repoPath);
-            pushToRepository(githubRepository, gitHubUser, repoPath, accessToken.getAccessToken());
+            GitHubUtil.initAndPushToGitHubRepository(
+                    githubRepository, gitHubUser, repoPath, accessToken.getAccessToken());
 
             if (LOG.isDebugEnabled()) {
                 LOG.debug("Successfully pushed application to " + githubRepository);
@@ -202,56 +195,6 @@ public class GitHubCreateService extends AbstractCreateController {
                     generatorContext);
         } catch (Exception e) {
             LOG.error("Error generating application: " + e.getMessage(), e);
-            throw new IOException(e.getMessage(), e);
-        }
-    }
-
-    /**
-     * Push files in {@code localPath} to the GitHub repository {@code repository}.
-     *
-     * @param repository github repository
-     * @param user       github user
-     * @param localPath  local directry with generated content
-     * @param token      github user access token
-     * @throws IOException
-     */
-    protected void pushToRepository(
-            @NotNull GitHubRepository repository,
-            @NotNull GitHubUser user,
-            @NotNull Path localPath,
-            @NotNull String token) throws IOException {
-        try {
-            final String name = StringUtils.isEmpty(user.getName()) ? user.getLogin() : user.getName();
-            final String email = StringUtils.isEmpty(user.getEmail()) ? user.getLogin() : user.getEmail();
-
-            Git gitRepo = Git.init().setDirectory(localPath.toFile())
-                    .call();
-            gitRepo.add()
-                    .addFilepattern(".")
-                    .call();
-
-            gitRepo.commit()
-                    .setMessage(COMMIT_MESSAGE)
-                    .setAuthor(name, email)
-                    .setCommitter(name, email)
-                    .setSign(false).call();
-
-            Iterable<PushResult> pushResults = gitRepo.push()
-                    .setRemote(repository.getCloneUrl())
-                    .setCredentialsProvider(new UsernamePasswordCredentialsProvider(user.getLogin(), token))
-                    .call();
-
-            List<RemoteRefUpdate> failedRefUpdates = StreamSupport.stream(pushResults.spliterator(), false)
-                    .flatMap(pushResult -> pushResult.getRemoteUpdates().stream())
-                    .filter(remoteRefUpdate -> remoteRefUpdate.getStatus() != RemoteRefUpdate.Status.OK)
-                    .collect(Collectors.toList());
-
-            if (!failedRefUpdates.isEmpty()) {
-                String msg = String.format("Failed to push to %s repository.", repository.getName());
-                LOG.error(msg);
-                throw new IOException(msg);
-            }
-        } catch (GitAPIException e) {
             throw new IOException(e.getMessage(), e);
         }
     }
