+++ starter-core/src/main/java/io/micronaut/starter/client/github/oauth/GitHubOAuthOperations.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.starter.api.create.github.client.oauth;
+package io.micronaut.starter.client.github.oauth;
 
 import io.micronaut.http.annotation.Post;
 import io.micronaut.http.annotation.QueryValue;
