+++ starter-core/src/main/java/io/micronaut/starter/client/github/v3/GitHubSecret.java
@@ -13,53 +13,43 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.starter.test.github.client;
+package io.micronaut.starter.client.github.v3;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.micronaut.core.annotation.Introspected;
 
 /**
- * GitHub User.
- *
  * @author Pavol Gressa
- * @see <a href="https://docs.github.com/en/rest/reference/users">Api reference users</a>
  * @since 2.2
  */
 @Introspected
-public class GitHubUser {
-    private final String login;
-    private final String email;
-    private final String name;
+public class GitHubSecret {
+    private final String encryptedValue;
+    private final String keyId;
 
     @JsonCreator
-    public GitHubUser(
-            @JsonProperty("login") String login,
-            @JsonProperty("email")  String email,
-            @JsonProperty("name") String name) {
-        this.login = login;
-        this.email = email;
-        this.name = name;
+    public GitHubSecret(@JsonProperty("encrypted_value") String encryptedValue,
+                        @JsonProperty("key_id") String keyId) {
+        this.encryptedValue = encryptedValue;
+        this.keyId = keyId;
     }
 
-    public String getLogin() {
-        return login;
+    @JsonProperty("encrypted_value")
+    public String getEncryptedValue() {
+        return encryptedValue;
     }
 
-    public String getEmail() {
-        return email;
-    }
-
-    public String getName() {
-        return name;
+    @JsonProperty("key_id")
+    public String getKeyId() {
+        return keyId;
     }
 
     @Override
     public String toString() {
-        return "GitHubUser{" +
-                "login='" + login + '\'' +
-                ", email='" + email + '\'' +
-                ", name='" + name + '\'' +
+        return "GitHubSecret{" +
+                "encryptedValue='" + encryptedValue + '\'' +
+                ", keyId='" + keyId + '\'' +
                 '}';
     }
 }
