+++ starter-core/src/main/java/io/micronaut/starter/client/github/v3/GitHubWorkflowRun.java
@@ -13,53 +13,52 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.starter.api.create.github.client.v3;
+package io.micronaut.starter.client.github.v3;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.micronaut.core.annotation.Introspected;
 
 /**
- * GitHub User.
+ * GitHub workflow run.
  *
  * @author Pavol Gressa
- * @see <a href="https://docs.github.com/en/rest/reference/users">Api reference users</a>
  * @since 2.2
  */
 @Introspected
-public class GitHubUser {
-    private final String login;
-    private final String email;
-    private final String name;
+public class GitHubWorkflowRun {
+
+    private final Long id;
+    private final String status;
+    private final String conclusion;
 
     @JsonCreator
-    public GitHubUser(
-            @JsonProperty("login") String login,
-            @JsonProperty("email")  String email,
-            @JsonProperty("name") String name) {
-        this.login = login;
-        this.email = email;
-        this.name = name;
+    public GitHubWorkflowRun(@JsonProperty("id") Long id,
+                             @JsonProperty("status") String status,
+                             @JsonProperty("conclusion") String conclusion) {
+        this.id = id;
+        this.status = status;
+        this.conclusion = conclusion;
     }
 
-    public String getLogin() {
-        return login;
+    public Long getId() {
+        return id;
     }
 
-    public String getEmail() {
-        return email;
+    public String getStatus() {
+        return status;
     }
 
-    public String getName() {
-        return name;
+    public String getConclusion() {
+        return conclusion;
     }
 
     @Override
     public String toString() {
-        return "GitHubUser{" +
-                "login='" + login + '\'' +
-                ", email='" + email + '\'' +
-                ", name='" + name + '\'' +
+        return "GitHubWorkflowRun{" +
+                "id=" + id +
+                ", status='" + status + '\'' +
+                ", conclusion='" + conclusion + '\'' +
                 '}';
     }
 }
